movie-recommendation-app/backend/requirements-python313.txt
# Core Python Dependencies
setuptools>=68.0.0
wheel>=0.41.0

# Django and Core Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-yasg==1.21.7

# Database - Python 3.13 compatible version
# Use psycopg2-binary with specific version for Python 3.13
psycopg2-binary==2.9.9

# Caching and Redis
redis==5.0.1
django-redis==5.4.0

# HTTP Requests
requests==2.31.0

# Environment Configuration
python-decouple==3.8

# Filtering
django-filter==23.5

# Production Server
gunicorn==21.2.0

# Static Files
whitenoise==6.6.0

# Deployment-specific dependencies
certifi>=2023.7.22
charset-normalizer>=3.2.0
idna>=3.4
urllib3>=1.26.16

# Additional deployment dependencies
packaging>=21.3
pyparsing>=3.0.9 